Migrate FdmlUtils tests to munit
